Applied Chord Testing
Still haven't fixed bug - Whenever Major+ or minor+ mode is chosen, there is a 90% chance that the program freezes.

Narrowed it down to the chordPrevention rules, as when they are commented out there is no issue. Not really sure why they don't work, tried a few different ways......

However I believe implementing this logic later on, within the romanNumeralList variable, should work, something along the lines of if V/V is selected, force the next chord to be V, and then check the chord after to ensure it isn't the same.

Almost there!